refactor: setting -> settings로 변경, 프래그먼트명 변경
