+++ app/src/main/res/layout/fragment_settings_private_policy.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.setting.SettingSourceFragment">
+    tools:context=".presentation.settings.SettingsSupportFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/cl_source_topbar"
+        android:id="@+id/cl_privatepolicy_topbar"
         android:layout_width="match_parent"
         android:layout_height="60dp"
         app:layout_constraintEnd_toEndOf="parent"
@@ -15,7 +15,7 @@
         app:layout_constraintTop_toTopOf="parent">
 
         <ImageView
-            android:id="@+id/iv_source_back"
+            android:id="@+id/iv_privatepolicy_back"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="16dp"
@@ -30,7 +30,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:fontFamily="@font/pretendard_bold"
-            android:text="오픈 소스 고지"
+            android:text="개인 정보 보호 및 보안 안내"
             android:textSize="16sp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
@@ -46,7 +46,7 @@
         android:background="#E7E7E7"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/cl_source_topbar" />
+        app:layout_constraintTop_toBottomOf="@id/cl_privatepolicy_topbar" />
 
     <WebView
         android:id="@+id/webview_support"
