+++ app/src/main/res/navigation/nav_graph.xml
@@ -79,8 +79,8 @@
 
     <fragment
         android:id="@+id/setting_support"
-        android:name="com.stopsmoke.kekkek.presentation.setting.SettingSupportFragment"
+        android:name="com.stopsmoke.kekkek.presentation.settings.SettingsPrivatePolicyFragment"
         app:route="setting_support"
-        tools:layout="@layout/fragment_setting_support"/>
+        tools:layout="@layout/fragment_settings_support"/>
 
 </navigation>
